only new table id datatype to BIGINT

+++ server/migrations/20190913075424-create-stripe-payments.js
@@ -32,7 +32,7 @@ module.exports = {
           type: "NUMERIC(10, 2)"
         },
         status: {
-          type: Sequelize.BIGINT
+          type: Sequelize.INTEGER
         },
         email: {
           type: Sequelize.STRING
